Backend/yolo-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
ultralytics>=8.0.200
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
